Move more code to Scala2
